dataflow-pipelines/flex-pubsub-dataflow-biqquery/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
--extra-index-url https://europe-west1-python.pkg.dev/gborel-sample-project/python-repo/simple/

annotated-types==0.7.0
    # via pydantic
apache-beam[gcp]==2.57.0
    # via
    #   -r requirements.in
    #   gborelpy
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
cachetools==5.4.0
    # via
    #   apache-beam
    #   google-auth
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
cloudpickle==2.2.1
    # via apache-beam
crcmod==1.7
    # via apache-beam
dill==0.3.1.1
    # via apache-beam
dnspython==2.6.1
    # via pymongo
docopt==0.6.2
    # via hdfs
docstring-parser==0.16
    # via google-cloud-aiplatform
fastavro==1.9.5
    # via apache-beam
fasteners==0.19
    # via
    #   apache-beam
    #   google-apitools
gborelpy==0.4.0
    # via -r requirements.in
google-api-core[grpc]==2.19.1
    # via
    #   -r requirements.in
    #   apache-beam
    #   gborelpy
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-bigtable
    #   google-cloud-core
    #   google-cloud-datastore
    #   google-cloud-dlp
    #   google-cloud-language
    #   google-cloud-pubsub
    #   google-cloud-pubsublite
    #   google-cloud-recommendations-ai
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-spanner
    #   google-cloud-storage
    #   google-cloud-videointelligence
    #   google-cloud-vision
google-apitools==0.5.31
    # via apache-beam
google-auth==2.32.0
    # via
    #   -r requirements.in
    #   apache-beam
    #   google-api-core
    #   google-auth-httplib2
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-dlp
    #   google-cloud-language
    #   google-cloud-pubsub
    #   google-cloud-recommendations-ai
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-storage
    #   google-cloud-videointelligence
    #   google-cloud-vision
google-auth-httplib2==0.2.0
    # via apache-beam
google-cloud-aiplatform==1.59.0
    # via apache-beam
google-cloud-bigquery==3.25.0
    # via
    #   apache-beam
    #   google-cloud-aiplatform
google-cloud-bigquery-storage==2.25.0
    # via apache-beam
google-cloud-bigtable==2.24.0
    # via apache-beam
google-cloud-core==2.4.1
    # via
    #   apache-beam
    #   google-cloud-bigquery
    #   google-cloud-bigtable
    #   google-cloud-datastore
    #   google-cloud-spanner
    #   google-cloud-storage
google-cloud-datastore==2.19.0
    # via apache-beam
google-cloud-dlp==3.18.1
    # via apache-beam
google-cloud-language==2.13.4
    # via apache-beam
google-cloud-pubsub==2.22.0
    # via
    #   -r requirements.in
    #   apache-beam
    #   google-cloud-pubsublite
google-cloud-pubsublite==1.11.0
    # via apache-beam
google-cloud-recommendations-ai==0.10.11
    # via apache-beam
google-cloud-resource-manager==1.12.4
    # via google-cloud-aiplatform
google-cloud-secret-manager==2.20.0
    # via -r requirements.in
google-cloud-spanner==3.47.0
    # via apache-beam
google-cloud-storage==2.17.0
    # via
    #   -r requirements.in
    #   apache-beam
    #   google-cloud-aiplatform
google-cloud-videointelligence==2.13.4
    # via apache-beam
google-cloud-vision==3.7.3
    # via apache-beam
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.63.2
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.1
    # via
    #   google-cloud-bigtable
    #   google-cloud-pubsub
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-spanner
grpc-interceptor==0.15.4
    # via google-cloud-spanner
grpcio==1.65.1
    # via
    #   apache-beam
    #   google-api-core
    #   google-cloud-pubsub
    #   google-cloud-pubsublite
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpc-interceptor
    #   grpcio-status
grpcio-status==1.62.2
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   google-cloud-pubsublite
hdfs==2.7.3
    # via apache-beam
httplib2==0.22.0
    # via
    #   apache-beam
    #   google-apitools
    #   google-auth-httplib2
    #   oauth2client
idna==3.7
    # via requests
js2py==0.74
    # via apache-beam
jsonpickle==3.2.2
    # via apache-beam
jsonschema==4.23.0
    # via apache-beam
jsonschema-specifications==2023.12.1
    # via jsonschema
numpy==1.26.4
    # via
    #   apache-beam
    #   pyarrow
    #   shapely
oauth2client==4.1.3
    # via google-apitools
objsize==0.7.0
    # via apache-beam
orjson==3.10.6
    # via apache-beam
overrides==7.7.0
    # via google-cloud-pubsublite
packaging==24.1
    # via
    #   apache-beam
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
proto-plus==1.24.0
    # via
    #   apache-beam
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery-storage
    #   google-cloud-bigtable
    #   google-cloud-datastore
    #   google-cloud-dlp
    #   google-cloud-language
    #   google-cloud-pubsub
    #   google-cloud-recommendations-ai
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-spanner
    #   google-cloud-videointelligence
    #   google-cloud-vision
protobuf==4.25.3
    # via
    #   apache-beam
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery-storage
    #   google-cloud-bigtable
    #   google-cloud-datastore
    #   google-cloud-dlp
    #   google-cloud-language
    #   google-cloud-pubsub
    #   google-cloud-recommendations-ai
    #   google-cloud-resource-manager
    #   google-cloud-secret-manager
    #   google-cloud-spanner
    #   google-cloud-videointelligence
    #   google-cloud-vision
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==16.1.0
    # via apache-beam
pyarrow-hotfix==0.6
    # via apache-beam
pyasn1==0.6.0
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via
    #   google-auth
    #   oauth2client
pydantic==2.8.2
    # via google-cloud-aiplatform
pydantic-core==2.20.1
    # via pydantic
pydot==1.4.2
    # via apache-beam
pyjsparser==2.7.1
    # via js2py
pymongo==4.8.0
    # via apache-beam
pyparsing==3.1.2
    # via
    #   httplib2
    #   pydot
python-dateutil==2.9.0.post0
    # via
    #   apache-beam
    #   google-cloud-bigquery
pytz==2024.1
    # via apache-beam
redis==5.0.7
    # via apache-beam
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.15
    # via apache-beam
requests==2.31.0
    # via
    #   apache-beam
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   hdfs
rpds-py==0.19.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
shapely==2.0.5
    # via google-cloud-aiplatform
six==1.16.0
    # via
    #   google-apitools
    #   hdfs
    #   js2py
    #   oauth2client
    #   python-dateutil
sqlparse==0.5.1
    # via google-cloud-spanner
typing-extensions==4.12.2
    # via
    #   apache-beam
    #   pydantic
    #   pydantic-core
tzlocal==5.2
    # via js2py
urllib3==2.2.2
    # via requests
zstandard==0.23.0
    # via apache-beam
